Creating stage directory if not there

+++ scripts/up.sh
@@ -3,15 +3,20 @@
 CURR_DIR="$( cd "$( dirname "${BASH_SOURCE[0]}" )" && pwd )"
 STAGE_DIR="$CURR_DIR/../stage"
 
-echo "This script will delete the contents of $STAGE_DIR, would you like to continue?"
-select yn in "Yes" "No"; do
-    case $yn in
-        Yes ) break;;
-        No ) exit;;
-    esac
-done
+if [ ! -d "$STAGE_DIR" ]; then
+  echo "Creating stage directory..."
+  mkdir $STAGE_DIR
+else
+  echo "This script will delete the contents of $STAGE_DIR, would you like to continue?"
+  select yn in "Yes" "No"; do
+      case $yn in
+          Yes ) break;;
+          No ) exit;;
+      esac
+  done
 
-rm -rf $STAGE_DIR/*
+  rm -rf $STAGE_DIR/*
+fi
 
 # Copy the latest artifact to the stage directory and extract it
 $CURR_DIR/stage.sh
